Don't copy grootfs in with the rest of garden-runc-release
On our 17404 windows concourse worker, tar.exe is now included in two
places:

  1. c:\var\vcap\bosh\bin
  2. c:\windows\system32

When concourse tries to stream garden-runc-release onto the windows
worker, it shells out to tar and that tar fails.

There was a previous PR to concourse-windows-worker to remove
c:\var\vcap\bosh\bin from the PATH which prevented a similar issue for
the Diego team (https://github.com/pivotal-cf-experimental/concourse-windows-worker-release/pull/5)

This PR does not work on 17404, since tar is now always on the PATH

Since grootfs is not needed for us to run GATS, and since removing
allows the stream in to succeed, we can work around this for now.

[#153833937]

Signed-off-by: Sam Smith <sesmith177@gmail.com>