Set TMP, TEMP, and GOTMPDIR in our task files
These seem to all be required.
Go build places intermediate files in GOTMPDIR if set since 1.10.

We tried setting only one variable through the pipeline and setting the
others through Powershell, but it didn't work for some reason that we
don't understand.

We will add a job to our windows-tools-release to setup the new TMP dir
when the worker comes up.

[#161954157]

Signed-off-by: Natalie Arellano <narellano@pivotal.io>