Add winc-develop to windowsfs pipeline.
- after cert-injector stuff is in a release of winc that is in the
release candidate of cd-deployment, then we can undo this work and go
back to pulling the official release of winc. We want to isolate the
variable we are testing in this pipeline, i.e. the develop branch of
windowsfs, and if we continue to use the develop branch of winc, then we
introduce two changing variables in this pipeline and it will make it
harder to debug.

Signed-off-by: Genevieve LEsperance <glesperance@pivotal.io>