Scale up number of routers to 2 for alfredo and pesto
If there is 1 instance of the router, there is a 502 error that occurs
frequently but not every time. If there are 3 instances of the router,
there is an issue which *looks like* bbl would not create backend-services
for more than 2 zones.
See github.com/cloudfoundry/bosh-bootloader/blob/v8.1.1/terraform/gcp/template_generator.go#L56-L58

Signed-off-by: Arjun Sreedharan <asreedharan@pivotal.io>