Winc pipeline: Move cats to gcp pesto
This commit partially reverts 60d8e44 but we keep using
garden-windows-environments to get vars.yml

It doesn't look like bbl on Azure supports cf-ssh out of the
box. We will have to make changes to terraform files and maintain
it if we want to use Azure.

See:
- https://github.com/cloudfoundry/bosh-bootloader/issues/377#issuecomment-425994751
- https://cloudfoundry.slack.com/archives/C0FA9UNAH/p1565361584020900

If/when we get to know Azure can be used easily, we can revisit this problem.

[#161892048] (https://www.pivotaltracker.com/story/show/161892048)

Signed-off-by: Arjun Sreedharan <asreedharan@pivotal.io>